agregar tarjeta persisistencia
duda entre usar OneToOne o embebled con PostPaga y Tarjeta

+++ TallerJava2024/src/main/java/org/tallerjava/moduloMediosDePago/aplicacion/impl/servicioMediosDePagoImpl.java
@@ -27,7 +27,7 @@ public class servicioMediosDePagoImpl implements servicioMediosDePago {
 	 private ServicioPagoFacade servicioPagoFacade;
 
 	@Override
-	public void altaCliente(long ci, int nroTarjeta, LocalDateTime fechaVtoTarjeta, String nombreCompletoUsuario) {
+	public void altaCliente(int ci, int nroTarjeta, LocalDateTime fechaVtoTarjeta, String nombreCompletoUsuario) {
 		//LLAMA AL MODULO GESTION CLIENTE TELEPEAJE CON CI DEBE EXISTIR PREVIAMENTE
 		//HAY QUE GUARDARLO AQUI?
 		servicioPagoFacade.agregarTarjeta(ci, nroTarjeta, fechaVtoTarjeta, nombreCompletoUsuario);
