+++ TallerJava2024/src/main/java/org/tallerjava/moduloMediosDePago/aplicacion/servicioMediosDePago.java
@@ -8,7 +8,7 @@ import org.tallerjava.moduloMediosDePago.dominio.*;
 
 public interface servicioMediosDePago {
 	
-	public void altaCliente(long ci, int nroTarjeta, LocalDateTime fechaVtoTarjeta, String nombreCompletoUsuario); //Da de alta al cliente
+	public void altaCliente(int ci, int nroTarjeta, LocalDateTime fechaVtoTarjeta, String nombreCompletoUsuario); //Da de alta al cliente
 	public int notificarPago(Cliente cliente, Vehiculo vehiculo, double importe, Tarjeta tarjeta); //notifica del pago al Sistema externo de Medios De Pago
 	public List<Pago> consultaDePagos(Date fechaInicial, Date fechaFinal); //importe total de pagos realizados agrupados  por día
 	public List<Pago> consultaDePagos(Cliente cliente); //devuelve los pagos realizados por un Cliente en particular
